Throw errors when certain DB operations fail

+++ src/services/posts_service.js
@@ -1,9 +1,14 @@
 const Post = require("../models/post");
 const Comment = require("../models/comment");
+const AppError = require("../utils/app_error");
 
 class PostsService {
    async getAllPosts() {
-      return await Post.find().populate("author").exec();
+      try {
+         await Post.find().populate("author").exec();
+      } catch (error) {
+         new AppError("Failed to retrieve posts");
+      }
    }
 
    async getPost(postId) {
@@ -12,6 +17,10 @@ class PostsService {
          Comment.find({ post_id: postId }).populate("author").exec(),
       ]);
 
+      if (post === null) {
+         throw new AppError("Post not found", 404);
+      }
+
       return {
          post,
          comments,
@@ -19,7 +28,11 @@ class PostsService {
    }
 
    async getPostComments(postId) {
-      return await Comment.find({ post_id: postId }).exec();
+      const comments = await Comment.find({ post_id: postId }).exec();
+
+      if (comments.length === 0) {
+         throw new AppError("No comments were found for this post", 404);
+      }
    }
 
    async createPost(authorId, title, text, isPublished) {
@@ -38,17 +51,29 @@ class PostsService {
    }
 
    async updatePost(authorId, title, text, isPublished, postId) {
-      const post = {
+      const post = await findById(postId).exec();
+
+      if (post === null) {
+         throw new AppError("Invalid resource identifier");
+      }
+
+      const updatedPost = {
          author: authorId,
          title,
          text,
          is_published: isPublished,
       };
 
-      await Post.findByIdAndUpdate(postId, { $set: post }).exec();
+      await Post.findByIdAndUpdate(postId, { $set: updatedPost }, { runValidators: true }).exec();
    }
 
    async deletePost(postId) {
+      const post = await findById(postId).exec();
+
+      if (post === null) {
+         throw new AppError("Invalid resource identifier");
+      }
+
       await Post.findByIdAndDelete(postId).exec();
    }
 }
